requirements.txt
python-dotenv
google-generativeai
streamlit 
drug-named-entity-recognition